Add files via upload
DOHMH New York City Restaurant Inspection Results dataset 
because it has a wealth of information it provides about the restaurant industry in 
New York City. It contains data on more than 200,000 restaurant inspections 
conducted over more than 10 years, providing insights into the health and safety of 
the city's food service establishments.
But I have analyzed that the dataset has many rows and columns. To achieve the best 
results, let's narrow down the dataset to only include the year "2021" (just after the 
end of the COVID pandemic) and analyze how many restaurant violations are still 
critical in New York.
After filtering the data frame to keep only 
records for 2021, I have taken the required columns which are the violation 
description and critical flag. Also, designed a model which can predict whether any 
kind of violation is critical or not within the system